refactor(bases): extract render cache to break circular require
Move clearRenderCache into basesRenderCache so linkUtils can clear the cache without requiring basesPlugin (which requires linkUtils).
